Allow multiple instances of GenericPluginManager
Summary:
This is necessary as several of our applications are using this, which
then fails when all are running inside the same process in Kontact.

This fixes all applications inside Kontact seeing the same set of tools
plugins, which of course don't work for all but one of them.

Reviewers: #kde_pim, mlaurent

Reviewed By: #kde_pim, mlaurent

Subscribers: mlaurent

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D7429