Fix pimcommon CMake
The configuration file for pimcommon was actually missing KIOCore, and
at the same time it worked by accident by linking QtNetwork, which
wasn't a public link target.

This change adds the KIOCore dependency where it belongs and fixes the
link targets properly by adding the actual dependencies (Qt Core, Gui,
Widgets, DBus). Qt Network is a private link target, so it is moved
where it belongs.

Reviewed and ack'ed by krop on IRC.